+++ app/src/main/java/com/github/se/studybuddies/ui/video_call/VideoCallScreen.kt
@@ -1,6 +1,5 @@
 package com.github.se.studybuddies.ui.video_call
 
-import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -8,7 +7,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -17,6 +15,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
+import com.github.se.studybuddies.viewModels.VideoCallViewModel
 import io.getstream.video.android.compose.theme.VideoTheme
 import io.getstream.video.android.compose.ui.components.call.activecall.CallContent
 import io.getstream.video.android.compose.ui.components.call.controls.ControlActions
@@ -28,21 +27,19 @@ import io.getstream.video.android.compose.ui.components.call.renderer.LayoutType
 import io.getstream.video.android.compose.ui.components.call.renderer.ParticipantVideo
 import io.getstream.video.android.compose.ui.components.call.renderer.ParticipantsLayout
 import io.getstream.video.android.compose.ui.components.call.renderer.RegularVideoRendererStyle
-import io.getstream.video.android.core.Call
-import io.getstream.video.android.core.StreamVideo
 
 // Design UI elements using Jetpack Compose
 @Composable
 fun VideoCallScreen(
-    call: Call,
+    videoCallVM: VideoCallViewModel,
     keepCallConnected: () -> Unit = {},
     onCallDisconnected: () -> Unit = {}
 ) {
-  val isCameraEnabled by call.camera.isEnabled.collectAsState()
-  val isMicrophoneEnabled by call.microphone.isEnabled.collectAsState()
+  val isCameraEnabled by videoCallVM.call.camera.isEnabled.collectAsState()
+  val isMicrophoneEnabled by videoCallVM.call.microphone.isEnabled.collectAsState()
   val layout by remember { mutableStateOf(LayoutType.DYNAMIC) }
 
-  JoinCall(call)
+  videoCallVM.JoinCall()
 
   VideoTheme {
     Box(modifier = Modifier.fillMaxSize().testTag("video_call_screen")) {
@@ -51,13 +48,13 @@ fun VideoCallScreen(
               Modifier.fillMaxSize()
                   .background(color = VideoTheme.colors.appBackground)
                   .testTag("call_content"),
-          call = call,
+          call = videoCallVM.call,
           enableInPictureInPicture = true,
           layout = layout,
           onBackPressed = { keepCallConnected.invoke() },
           videoContent = {
             ParticipantsLayout(
-                call = call,
+                call = videoCallVM.call,
                 modifier =
                     Modifier.fillMaxSize()
                         .weight(1f)
@@ -67,7 +64,7 @@ fun VideoCallScreen(
                 videoRenderer = { modifier, _, participant, style ->
                   ParticipantVideo(
                       modifier = modifier.padding(4.dp).clip(RoundedCornerShape(8.dp)),
-                      call = call,
+                      call = videoCallVM.call,
                       participant = participant,
                       style = style,
                   )
@@ -76,7 +73,7 @@ fun VideoCallScreen(
           },
           controlsContent = {
             ControlActions(
-                call = call,
+                call = videoCallVM.call,
                 modifier = Modifier.testTag("control_actions"),
                 actions =
                     listOf(
@@ -84,18 +81,18 @@ fun VideoCallScreen(
                           ToggleCameraAction(
                               modifier = Modifier.size(52.dp),
                               isCameraEnabled = isCameraEnabled,
-                              onCallAction = { toggleCamera(call, it.isEnabled) })
+                              onCallAction = { videoCallVM.enableCamera(it.isEnabled) })
                         },
                         {
                           ToggleMicrophoneAction(
                               modifier = Modifier.size(52.dp),
                               isMicrophoneEnabled = isMicrophoneEnabled,
-                              onCallAction = { toggleMicrophone(call, it.isEnabled) })
+                              onCallAction = { videoCallVM.enableMicrophone(it.isEnabled) })
                         },
                         {
                           FlipCameraAction(
                               modifier = Modifier.size(52.dp),
-                              onCallAction = { call.camera.flip() })
+                              onCallAction = { videoCallVM.call.camera.flip() })
                         },
                         {
                           LeaveCallAction(
@@ -104,8 +101,8 @@ fun VideoCallScreen(
                                       VideoTheme.dimens.controlActionsButtonSize,
                                   ),
                               onCallAction = {
-                                removeActiveCall()
-                                leaveCall(call)
+                                videoCallVM.removeActiveCall()
+                                videoCallVM.leaveCall()
                                 onCallDisconnected.invoke()
                               },
                           )
@@ -115,46 +112,4 @@ fun VideoCallScreen(
       )
     }
   }
-}
-
-@Composable
-private fun JoinCall(call: Call) {
-  LaunchedEffect(call) {
-    if (StreamVideo.instance().state.activeCall.value == call) {
-      Log.d("MyPrint", "Active call is the same as the call we are trying to join")
-    } else {
-      try {
-        call.join()
-        Log.d("MyPrint", "Trying to join call")
-      } catch (e: Exception) {
-        Log.d("MyPrint", "Trying to join call got exception, leave call")
-        call.leave()
-        call.join()
-      } finally {
-        Log.d("MyPrint", "Keeping active call")
-        keepActiveCall(call)
-      }
-    }
-  }
-}
-
-fun leaveCall(call: Call) {
-  Log.d("MyPrint", "Trying to leave call")
-  call.leave()
-}
-
-fun keepActiveCall(call: Call) {
-  StreamVideo.instance().state.setActiveCall(call)
-}
-
-fun removeActiveCall() {
-  StreamVideo.instance().state.removeActiveCall()
-}
-
-fun toggleCamera(call: Call, enable: Boolean) {
-  call.camera.setEnabled(enable)
-}
-
-fun toggleMicrophone(call: Call, enable: Boolean) {
-  call.microphone.setEnabled(enable)
 }
