+++ app/src/main/java/com/github/se/studybuddies/viewModels/VideoCallViewModel.kt
@@ -1,23 +1,22 @@
 package com.github.se.studybuddies.viewModels
 
 import android.util.Log
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import io.getstream.video.android.core.Call
 import io.getstream.video.android.core.DeviceStatus
 import io.getstream.video.android.core.StreamVideo
 import javax.inject.Inject
-import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
 
-class CallLobbyViewModel @Inject constructor(val uid: String, val callType: String) : ViewModel() {
-
+class VideoCallViewModel @Inject constructor(val uid: String, val call: Call) : ViewModel() {
   data class UiState(
       val isLoading: Boolean = true,
       val isCameraEnabled: Boolean =
@@ -25,44 +24,42 @@ class CallLobbyViewModel @Inject constructor(val uid: String, val callType: Stri
       val isMicrophoneEnabled: Boolean = false
   )
 
-  val call: Call by lazy {
-    val streamVideo = StreamVideo.instance()
-    val call = streamVideo.call(callType, uid)
-    viewModelScope.launch {
-      // create the call if it doesn't exist - this will also load the settings for the call,
-      // this way the lobby screen can already display the right mic/camera settings
-      // This also starts listening to the call events to get the participant count
-      val callGetOrCreateResult = call.create()
-      if (callGetOrCreateResult.isFailure) {
-        // in demo we can ignore this. The lobby screen will just display default camera/video,
-        // but we will show an error
-        Log.e(
-            "CallLobbyViewModel",
-            "Failed to create the call ${callGetOrCreateResult.errorOrNull()}",
-        )
-        _event.emit(CallLobbyEvent.JoinFailed(callGetOrCreateResult.errorOrNull()?.message))
-      }
-    }
-    call
-  }
-
   val state: StateFlow<UiState> =
       flow {
-            emit(UiState(isLoading = true))
+            emit(VideoCallViewModel.UiState(isLoading = true))
             val isCameraEnabled = call.camera.status.first() is DeviceStatus.Enabled
             val isMicrophoneEnabled = call.microphone.status.first() is DeviceStatus.Enabled
             emit(
-                UiState(
+                VideoCallViewModel.UiState(
                     isLoading = false,
                     isCameraEnabled = isCameraEnabled,
                     isMicrophoneEnabled = isMicrophoneEnabled))
           }
-          .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), UiState())
-
+          .stateIn(
+              viewModelScope, SharingStarted.WhileSubscribed(5000), VideoCallViewModel.UiState())
   private val _isLoading: MutableStateFlow<Boolean> = MutableStateFlow(false)
   internal val isLoading: StateFlow<Boolean> = _isLoading
 
-  private val _event: MutableSharedFlow<CallLobbyEvent> = MutableSharedFlow()
+  @Composable
+  fun JoinCall() {
+    LaunchedEffect(call) {
+      if (StreamVideo.instance().state.activeCall.value == call) {
+        Log.d("MyPrint", "Active call is the same as the call we are trying to join")
+      } else {
+        try {
+          call.join()
+          Log.d("MyPrint", "Trying to join call")
+        } catch (e: Exception) {
+          Log.d("MyPrint", "Trying to join call got exception, leave call")
+          call.leave()
+          call.join()
+        } finally {
+          Log.d("MyPrint", "Keeping active call")
+          keepActiveCall(call)
+        }
+      }
+    }
+  }
 
   fun enableCamera(enabled: Boolean) {
     call.camera.setEnabled(enabled)
@@ -71,11 +68,17 @@ class CallLobbyViewModel @Inject constructor(val uid: String, val callType: Stri
   fun enableMicrophone(enabled: Boolean) {
     call.microphone.setEnabled(enabled)
   }
-}
 
-sealed interface CallLobbyEvent {
+  fun leaveCall() {
+    Log.d("MyPrint", "Trying to leave call")
+    call.leave()
+  }
 
-  data object JoinCall : CallLobbyEvent
+  private fun keepActiveCall(call: Call) {
+    StreamVideo.instance().state.setActiveCall(call)
+  }
 
-  data class JoinFailed(val reason: String?) : CallLobbyEvent
+  fun removeActiveCall() {
+    StreamVideo.instance().state.removeActiveCall()
+  }
 }
